+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## Fixed
+### Added
+- Add `application_headers` to `Pigeon::emmit` as last parameter
+### Fixed
 - Auto declare exchange with `routing`
 - Fix `application_headers` to `AMQPTable`